tool3/requirements.txt
google-generativeai
openai 
anthropic 
python-dotenv